chore: refactor some enums
