+++ common/src/main/java/com/ryderbelserion/crazycrates/common/enums/types/FillerType.java
@@ -1,4 +1,4 @@
-package com.badbones69.crazycrates.api.enums;
+package com.ryderbelserion.crazycrates.common.enums.types;
 
 public enum FillerType {
