+++ paper/src/main/java/com/badbones69/crazycrates/listeners/crates/types/CosmicCrateListener.java
@@ -4,8 +4,8 @@ import com.badbones69.crazycrates.api.events.PlayerPrizeEvent;
 import com.badbones69.crazycrates.api.events.PlayerReceiveKeyEvent;
 import com.badbones69.crazycrates.api.builders.ItemBuilder;
 import com.ryderbelserion.crazycrates.common.plugin.util.Methods;
-import com.badbones69.crazycrates.managers.events.EventManager;
-import com.badbones69.crazycrates.managers.events.enums.EventType;
+import com.badbones69.crazycrates.managers.EventManager;
+import com.ryderbelserion.crazycrates.common.enums.types.EventType;
 import com.ryderbelserion.vital.paper.api.enums.Support;
 import com.ryderbelserion.vital.common.util.AdvUtil;
 import io.papermc.paper.persistence.PersistentDataContainerView;
