+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/CrateManager.java
@@ -3,7 +3,7 @@ package com.badbones69.crazycrates.tasks.crates;
 import ch.jalu.configme.SettingsManager;
 import com.Zrips.CMI.Modules.ModuleHandling.CMIModule;
 import com.badbones69.crazycrates.api.builders.CrateBuilder;
-import com.badbones69.crazycrates.managers.events.enums.EventType;
+import com.ryderbelserion.crazycrates.common.enums.types.EventType;
 import com.badbones69.crazycrates.tasks.menus.CrateMainMenu;
 import com.badbones69.crazycrates.api.objects.crates.CrateHologram;
 import com.badbones69.crazycrates.api.objects.crates.quadcrates.CrateSchematic;
@@ -559,10 +559,6 @@ public class CrateManager {
      * @param location the location that may be needed for some crate types.
      * @param checkHand if it just checks the players hand or if it checks their inventory.
      * @param eventType {@link EventType}
-     * @param player the player that is having the crate opened for them
-     * @param crate the crate that is being used
-     * @param location the location that may be needed for some crate types
-     * @param checkHand if it just checks the players hand or if it checks their inventory
      */
     public void openCrate(@NotNull final Player player, @NotNull final Crate crate, @NotNull final KeyType keyType, @NotNull final Location location, final boolean virtualCrate, final boolean checkHand, final EventType eventType) {
         openCrate(player, crate, keyType, location, virtualCrate, checkHand, false, eventType);
