+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/types/QuadCrate.java
@@ -3,7 +3,7 @@ package com.badbones69.crazycrates.tasks.crates.types;
 import com.badbones69.crazycrates.api.objects.crates.quadcrates.CrateSchematic;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.crates.CrateLocation;
-import com.badbones69.crazycrates.managers.events.enums.EventType;
+import com.ryderbelserion.crazycrates.common.enums.types.EventType;
 import com.ryderbelserion.vital.paper.util.structures.StructureManager;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
